// src/server/spider-client/rapport/spider-rapport/public/secondaire-general/enquete-rapide-de-entree/interfaces.ts
//Convert from clipboard (Shift + Ctrl + Alt + V)

export interface IChp1 {
  orderby: number;
  label: string;
  _6e: number;
  _5e: number;
  _4e: number;
  _3e: number;
  ST1: number;
  _2ndA: number;
  _2ndC: number;
  _1ereA: number;
  _1ereC: number;
  _1ereD: number;
  _TleA: number;
  _TleC: number;
  _TleD: number;
  ST2: number;
  TOTAL: number;
}

export interface IChp2 {
  orderby: number;
  label: string;
  _6e: number;
  _5e: number;
  _4e: number;
  _3e: number;
  ST1: number;
  _2ndA: number;
  _2ndC: number;
  _1ereA: number;
  _1ereC: number;
  _1ereD: number;
  _TleA: number;
  _TleC: number;
  _TleD: number;
  ST2: number;
  TOTAL: number;
}

export interface IChp3 {
  orderby: number;
  label: string;
  maths: number;
  pc: number;
  svt: number;
  fr: number;
  ang: number;
  all: number;
  esp: number;
  philo: number;
  hg: number;
  mus: number;
  ap: number;
  eps: number;
  hg_lm: number;
  lm_edhc: number;
  ang_eps: number;
  math_tice: number;
  pc_svt: number;
  svt_pc: number;
  total: number;
}

export interface IChp4 {
  orderby: number;
  label: string;
  chefetab: number;
  adjoint_chefetab: number;
  indend: number;
  indend_adjoint: number;
  insp_educ: number;
  insp_ori: number;
  educ: number;
  total: string;
}

export interface IChp5 {
  orderby: number;
  label: string;
  secret: number;
  info: number;
  laboratin: number;
  biblio: number;
  reprog: number;
  infi: number;
  man: number;
  archiv: number;
  gardi: number;
  total: number;
}// 
//fin